Organized some components
- Changed some images of carousel
- Deleted caption01, caption02 and caption03 component

https://trello.com/c/qFJIuyqY/66-conectar-aplicacion-con-back-end

We worked together: afforeroc capinillab lemartinp
