+++ src/app/app.router.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { HomeComponent } from './home/home.component';
 import { ContactComponent } from './contact/contact.component';
-import { ProducersComponent } from './producers/producers.component';
+import { ProducerComponent } from './producer/producer.component';
 import { WarehouseComponent } from './warehouse/warehouse.component';
 import { AccountComponent } from './account/account.component';
 import { PlaceComponent } from './place/place.component';
@@ -13,7 +13,7 @@ export const router: Routes = [
   { path: '', redirectTo: 'home', pathMatch: 'full' },
   { path: 'home', component: HomeComponent },
   { path: 'contact', component: ContactComponent },
-  { path: 'producers', component: ProducersComponent },
+  { path: 'producer', component: ProducerComponent },
   { path: 'warehouse', component: WarehouseComponent },
   { path: 'account', component: AccountComponent },
   { path: 'place', component: PlaceComponent },
