--- src/app/caption01/caption01.component.html
@@ -1,15 +1,0 @@
-<h1>Where you can be a part<br>
-  of the farming market!</h1>
-  <br />
-  <h3>Give it a try and pick one these areas<br>
-    that you can be a part of:</h3>
-  <p>
-    <span class="glyphicon glyphicon-search" aria-hidden="true"></span>
-    Producer: Find an easyway to distribute your products. <br>
-    <span class="glyphicon glyphicon-globe" aria-hidden="true"></span>
-    Distributor or Warehouse: Get new source of products. <br>
-    <span class="glyphicon glyphicon-qrcode" aria-hidden="true"></span>
-    Client: Get information about the products. <br>
-  </p>
-  <br />
-  <p>If you aren't a part of it yet go ahead!</p>
