--- src/app/caption01/caption01.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-caption01',
-  templateUrl: './caption01.component.html',
-  styleUrls: ['./caption01.component.css']
-})
-export class Caption01Component implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
