--- src/app/caption02/caption02.component.html
@@ -1,35 +1,0 @@
-<h1>About Producer</h1>
-<br />
-<ul>
-	<li>
-		<p> Know the minimal standards that you must<br>
-			accomplish in order to sell your products
-		</p>
-	</li>
-	<li>
-		<p> Be informed about where your proudcts are being distributed</p>
-	</li>
-	<li>
-		<p> Also get information about the conditions<br>
-			your products are being stored on
-		</p>
-	</li>
-</ul>
-
-<h1>About Distributor</h1>
-<br />
-<ul>
-	<li>
-		<p> Know the minimal standards that you must accomplish in <br>
-			order to store and distribute other peoples products
-		</p>
-	</li>
-	<li>
-		<p> Know everything you need about the products that you're <br>
-			buying to garantee a good quality to your customers
-		</p>
-	</li>
-	<li>
-		<p> You can get merchandise from a producer or another distributor!</p>
-	</li>
-</ul>
