--- src/app/caption02/caption02.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-caption02',
-  templateUrl: './caption02.component.html',
-  styleUrls: ['./caption02.component.css']
-})
-export class Caption02Component implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
