--- src/app/caption03/caption03.component.html
@@ -1,9 +1,0 @@
-<h1>Title 3<br>
-  About Client
-</h1>
-<br />
-<h3>Give it a try and pick one these areas<br>
-  that you can be a part of:
-</h3>
-<br />
-<p>If you aren't a part of it yet go ahead!</p>
