--- src/app/caption03/caption03.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-caption03',
-  templateUrl: './caption03.component.html',
-  styleUrls: ['./caption03.component.css']
-})
-export class Caption03Component implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
