+++ src/app/footer/footer.component.css
@@ -1,0 +1,1 @@
+
