+++ src/app/footer/footer.component.html
@@ -9,10 +9,7 @@
           <a routerLink="contact">Contact</a>
         </li>
         <li>
-          <a routerLink="producers">Producer</a>
-        </li>
-        <li>
-          <a routerLink="warehouse">Warehouse</a>
+          <a routerLink="producer">Producer</a>
         </li>
       </ul>
     </div>
