+++ src/app/header/header.component.css
@@ -2,5 +2,6 @@
 
 .navbar-brand{
   font-family: 'Pacifico', cursive;
-	font-size: 43px !important;
+	font-size: 28px !important;
 }
+
