+++ src/app/home/home.component.css
@@ -27,13 +27,3 @@ html,body{height:100%;}
 .carousel-inner {
   height: 100%;
 }
-
-.fill {
-  width: 100%;
-  height: 100%;
-  background-position: center;
-  -webkit-background-size: cover;
-  -moz-background-size: cover;
-  background-size: cover;
-  -o-background-size: cover;
-}
