+++ src/app/home/home.component.html
@@ -2,40 +2,77 @@
   <ol class="carousel-indicators">
     <li data-target="#amazing-carousel" data-slide-to="0" class="active"></li>
     <li data-target="#amazing-carousel" data-slide-to="1"></li>
-    <!--<li data-target="#amazing-carousel" data-slide-to="2"></li>-->
   </ol>
 
   <div class="carousel-inner" role="listbox">
 
     <div class="item active">
-      <img src="http://i42.tinypic.com/2wbykd5.png" alt="...">
-      <div class="container">
+      <img src="src/assets/banana.jpg" alt="...">
+      <div class="container-fluid">
         <div class="carousel-caption">
-          <app-caption01></app-caption01>
+          <h1>Where you can be a part<br>
+            of the farming market!</h1>
+          <br />
+          <h3>Give it a try and pick one these areas<br>
+            that you can be a part of:</h3>
+          <p>
+            <span class="glyphicon glyphicon-search" aria-hidden="true"></span>
+            Producer: Find an easyway to distribute your products. <br>
+            <span class="glyphicon glyphicon-globe" aria-hidden="true"></span>
+            Distributor or Warehouse: Get new source of products. <br>
+            <span class="glyphicon glyphicon-qrcode" aria-hidden="true"></span>
+            Client: Get information about the products. <br>
+          </p>
+          <br />
+          <p>If you aren't a part of it yet go ahead!</p>
           <a class = "btn btn-success" href="/account" role="button">Get start</a>
         </div>
    	  </div>
     </div>
 
     <div class="item">
-      <img src="http://i42.tinypic.com/2wbykd5.png" alt="...">
-      <div class="container">
+      <img src="src/assets/manzana.jpg" alt="...">
+      <div class="container-fluid">
         <div class="carousel-caption">
-          <app-caption02></app-caption02>
+          <h1>About Producer</h1>
+          <br />
+          <ul>
+            <li>
+              <p> Know the minimal standards that you must<br>
+                accomplish in order to sell your products
+              </p>
+            </li>
+            <li>
+              <p> Be informed about where your proudcts are being distributed</p>
+            </li>
+            <li>
+              <p> Also get information about the conditions<br>
+                your products are being stored on
+              </p>
+            </li>
+          </ul>
+
+          <h1>About Distributor</h1>
+          <br />
+          <ul>
+            <li>
+              <p> Know the minimal standards that you must accomplish in <br>
+                order to store and distribute other peoples products
+              </p>
+            </li>
+            <li>
+              <p> Know everything you need about the products that you're <br>
+                buying to garantee a good quality to your customers
+              </p>
+            </li>
+            <li>
+              <p> You can get merchandise from a producer or another distributor!</p>
+            </li>
+          </ul>
         </div>
       </div>
     </div>
-	<!--
-    <div class="item">
-      <img src="http://i42.tinypic.com/2wbykd5.png" alt="...">
-      <div class="container">
-        <div class="carousel-caption">
-          <app-caption03></app-caption03>
-          <p><button type="submit" class="btn btn-success btn-more-padding">Get Started</button></p>
-        </div>
-      </div>
-    </div>
-	//-->
+
     <a class="left carousel-control" href="#amazing-carousel" role="button" data-slide="prev">
       <span class="icon-prev" aria-hidden="true"></span>
       <span class="sr-only">Previous</span>
