+++ src/app/producer/producer.component.html
@@ -14,25 +14,25 @@
     <div class="tab-content col-md-5">
       <div id="getProducer" class="tab-pane fade in active">
         <div class="form">
-            <form #f="ngForm" (ngSubmit)="getProducer(f.value)">
-              <input type="text" id="inputID1" name="GetProducer" ngModel>
-              <button type="submit" class="btn btn-success">GET</button>
-              <h3>Your Information</h3>
-              <div class="form-group">
-                <label class="control-label col-sm-3" for="name">Name:</label>
-                <div class="col-sm-7">
-                  <p class="form-control-static" name="wName">{{data?.data.first_name}}</p>
-                </div>
-                <label class="control-label col-sm-3" for="lastname">Last Name:</label>
-                <div class="col-sm-7">
-                  <p class="form-control-static" name="wLastname">{{data?.data.last_name}}</p>
-                </div>
-                <label class="control-label col-sm-3" for="username">Username:</label>
-                <div class="col-sm-7">
-                  <p class="form-control-static" name="wUsername">{{data?.data.username}}</p>
-                </div>
+          <form #f="ngForm" (ngSubmit)="getProducer(f.value)">
+            <input type="text" id="inputID1" name="GetProducer" ngModel>
+            <button type="submit" class="btn btn-success">GET</button>
+            <h3>Your Information</h3>
+            <div class="form-group">
+              <label class="control-label col-sm-3" for="name">Name:</label>
+              <div class="col-sm-7">
+                <p class="form-control-static" name="wName">{{data?.data.first_name}}</p>
               </div>
-            </form>
+              <label class="control-label col-sm-3" for="lastname">Last Name:</label>
+              <div class="col-sm-7">
+                <p class="form-control-static" name="wLastname">{{data?.data.last_name}}</p>
+              </div>
+              <label class="control-label col-sm-3" for="username">Username:</label>
+              <div class="col-sm-7">
+                <p class="form-control-static" name="wUsername">{{data?.data.username}}</p>
+              </div>
+            </div>
+          </form>
         </div>
       </div>
 
@@ -50,11 +50,11 @@
             </div>
             <div class="form-group">
               <label for="pwd">Old Password:</label>
-  	          <input type="password" class="form-control" placeholder="Password">
+              <input type="password" class="form-control" placeholder="Password">
               <label for="pwd">New Password:</label>
-  	          <input type="password" class="form-control" placeholder="New Password">
+              <input type="password" class="form-control" placeholder="New Password">
               <label for="pwd">Confirm New Password:</label>
-  	          <input type="password" class="form-control" placeholder="Confirm New Password">
+              <input type="password" class="form-control" placeholder="Confirm New Password">
             </div>
             <div class="checkbox">
               <label><input type="checkbox">Are you sure?</label>
@@ -64,7 +64,7 @@
           </form>
         </div>
       </div>
-      <!--
+
       <div id="placesProducer" class="tap-pane fade">
         <app-place></app-place>
       </div>
@@ -72,7 +72,7 @@
       <div id="placesProduct" class="tap-pane fade">
         <app-product></app-product>
       </div>
-      -->
+
       <div id="deleteProducer" class="tap-pane fade">
         <h3>Delete Your Account</h3>
         <div class="form">
@@ -83,9 +83,9 @@
             </div>
             <div class="form-group">
               <label for="pwd">Password:</label>
-  	          <input type="password" class="form-control" placeholder="Password">
+              <input type="password" class="form-control" placeholder="Password">
               <label for="pwd">Confirm Password:</label>
-  	          <input type="password" class="form-control" placeholder="Confirm Password">
+              <input type="password" class="form-control" placeholder="Confirm Password">
             </div>
             <div class="checkbox">
               <label><input type="checkbox">Are you sure?</label>
